Split up STM32 pins files (#17212)
