+++ Marlin/src/pins/stm32f1/pins_BTT_SKR_MINI_E3_V1_0.h
