+++ Marlin/src/pins/stm32f1/pins_BTT_SKR_MINI_V1_1.h
