+++ Marlin/src/pins/stm32f1/pins_FYSETC_AIO_II.h
