+++ Marlin/src/pins/stm32f1/pins_GTM32_MINI_A30.h
