+++ Marlin/src/pins/stm32f1/pins_JGAURORA_A5S_A1.h
