+++ Marlin/src/pins/stm32f1/pins_STM3R_MINI.h
