+++ Marlin/src/pins/stm32f4/pins_FYSETC_S6.h
